incorporate new bezier

+++ include/BezierPattern.h
@@ -12,8 +12,8 @@ public:
     BezierPattern(const std::vector<BezierEnvelope>& envelopes);
     std::vector<uint16_t> getFrameAtTime(double time) const;
     std::vector<BezierEnvelope> envelopes;
-    double duration;
-    double numOutputs;
+    float duration;
+    uint8_t numOutputs;
 
 private:
 };
